extract table and gates
