+++ sha/src/circuit/config.rs
@@ -1,8 +1,9 @@
-use halo2_proofs::halo2curves::pairing::{Engine, MultiMillerLoop};
+use halo2_proofs::arithmetic::Field;
 use halo2_proofs::plonk::{Advice, Column, ConstraintSystem, Fixed, Instance, Selector};
 
 const ADVICE_NUM: usize = 4;
-const SELECTOR_NUM: usize = 3;
+const SELECTOR_NUM: usize = 1;
+const LOOKUP_SELECTOR_NUM: usize = 2;
 const FIXED_NUM: usize = 2;
 
 #[derive(Clone, Debug)]
@@ -10,11 +11,12 @@ pub struct ShaConfig {
     pub advices: [Column<Advice>; ADVICE_NUM],
     pub instance: Column<Instance>,
     pub selectors: [Selector; SELECTOR_NUM],
+    pub lookup_selectors: [Selector; LOOKUP_SELECTOR_NUM],
     pub fixed: [Column<Fixed>; FIXED_NUM],
 }
 
 impl ShaConfig {
-    pub fn new<E: MultiMillerLoop>(meta: &mut ConstraintSystem<<E as Engine>::Scalar>) -> Self {
+    pub fn new<F: Field>(meta: &mut ConstraintSystem<F>) -> Self {
         let advices = Self::create_columns::<ADVICE_NUM, _, _>(|| {
             let column = meta.advice_column();
             meta.enable_equality(column);
@@ -22,13 +24,16 @@ impl ShaConfig {
         });
         let instance = meta.instance_column();
         let selectors = Self::create_columns::<SELECTOR_NUM, _, _>(|| meta.selector());
+        let lookup_selectors =
+            Self::create_columns::<LOOKUP_SELECTOR_NUM, _, _>(|| meta.complex_selector());
         let fixed = Self::create_columns::<FIXED_NUM, _, _>(|| meta.fixed_column());
 
         Self {
-            advices: advices.try_into().unwrap(),
+            advices,
             instance,
             selectors,
-            fixed: fixed.try_into().unwrap(),
+            lookup_selectors,
+            fixed,
         }
     }
 
@@ -46,10 +51,10 @@ impl ShaConfig {
     }
 
     pub fn majority_selector(&self) -> Selector {
-        self.selectors[1]
+        self.lookup_selectors[0]
     }
 
     pub fn choose_selector(&self) -> Selector {
-        self.selectors[2]
+        self.lookup_selectors[1]
     }
 }
